qml logic

+++ sync/mysql.cpp
@@ -50,7 +50,20 @@ void MySql::initDB(const QString &database){
     db->setDatabaseName(d.absolutePath());
     if(db->open()){
         qyery = new QSqlQuery(*db);
-        QString qyer = QString("CREATE TABLE IF NOT EXISTS songs("
+
+        /*
+         *https://stackoverflow.com/questions/40863216/sqlite-why-is-foreign-key-constraint-not-working-here
+        */
+
+        QString qyer = QString("PRAGMA foreign_keys = ON");
+        if(!qyery->exec(qyer)){
+            sqlErrorLog(qyer);
+            throw InitDBError();
+            delete db;
+            return;
+        }
+
+        qyer = QString("CREATE TABLE IF NOT EXISTS songs("
                      "id INTEGER PRIMARY KEY AUTOINCREMENT, "
                      "name VARCHAR(100), "
                      "size INT NOT NULL, "
@@ -62,6 +75,7 @@ void MySql::initDB(const QString &database){
             delete db;
             return;
         }
+
 
         qyer = QString("CREATE UNIQUE INDEX IF NOT EXISTS isongs ON songs(name,size)");
         if(!qyery->exec(qyer)){
@@ -87,7 +101,7 @@ void MySql::initDB(const QString &database){
         qyer = QString("CREATE TABLE IF NOT EXISTS playlistsdata("
                      "playlist INT NOT NULL,"
                      "song INT NOT NULL,"
-                     "FOREIGN KEY(playlist) REFERENCES playlists(id)"
+                     "FOREIGN KEY(playlist) REFERENCES playlists(name)"
                         " ON UPDATE CASCADE"
                         " ON DELETE CASCADE,"
                      "FOREIGN KEY(song) REFERENCES songs(id)"
@@ -101,17 +115,19 @@ void MySql::initDB(const QString &database){
             return;
         }
 
-        qyer = QString("CREATE UNIQUE INDEX IF NOT EXISTS iplaylistsdata ON playlistsdata(playlist,song)");
+        qyer = QString("CREATE UNIQUE INDEX IF NOT EXISTS iplaylistsdata ON "
+                       "playlistsdata(playlist,song)");
         if(!qyery->exec(qyer)){
             sqlErrorLog(qyer);
             throw InitDBError();
             delete db;
             return;
         }
+
     }
 }
 
-void MySql::sqlErrorLog(const QString &qyery){
+void MySql::sqlErrorLog(const QString &qyery)const{
 #ifdef QT_DEBUG
             qDebug()<< qyery << ": fail:\n " <<this->qyery->lastError();
 #endif
@@ -153,8 +169,8 @@ int MySql::save(const Song &song){
     return result;
 }
 
-int MySql::save(const QString &url){
-    QFile f(url);
+int MySql::save(const QString &url){    
+    QFile f(QUrl(url).toLocalFile());
     if(!f.open(QIODevice::ReadOnly)){
         return -1;
     }
@@ -201,13 +217,15 @@ bool MySql::load(const SongHeader &song,Song &result){
     return true;
 }
 
-bool MySql::updateAvailableSongs(QList<SongHeader>& list, const QString& playList){
+bool MySql::updateAvailableSongs(QList<SongHeader>& list, const QString& playList, bool forEditing){
     QString qyer;
-    if(playList.isEmpty()){
+
+    if(playList.isEmpty() || playList == ALL_SONGS_LIST || forEditing){
         qyer = QString("SELECT id,name,size from songs");
     }else{
         qyer = QString("SELECT id,name,size from songs where "
-                       "id = (select song from playlistsdata where playlist='%0')").arg(playList);
+                       "id = (select song from playlistsdata where "
+                       "playlist='%0')").arg(playList);
     }
 
     if(!qyery->exec(qyer)){
@@ -219,10 +237,47 @@ bool MySql::updateAvailableSongs(QList<SongHeader>& list, const QString& playLis
 
     while(qyery->next()){
         SongHeader song;
+        song.isSelected = !forEditing || playList == ALL_SONGS_LIST;
         song.id = qyery->value(0).toInt();
         song.name = qyery->value(1).toString();
         song.size = qyery->value(2).toInt();
         list.push_back(song);
+    }
+
+    if(forEditing && list.size() > 0 && playList != ALL_SONGS_LIST){
+        QString qyer;
+        qyer = QString("select song from playlistsdata where "
+                           " playlist='%0'").arg(playList);
+
+        if(!qyery->exec(qyer)){
+            sqlErrorLog(qyer);
+            return false;
+        }
+
+        while(qyery->next()){
+            for(SongHeader& item:list){
+                int id = qyery->value(0).toInt();
+                if(item.id == id){
+                    item.isSelected = true;
+                    break;
+                }
+            }
+        }
+    }
+
+
+    return true;
+}
+
+bool MySql::updateAvailableSongs(QStringList& list, const QString& playList){
+
+    QList<SongHeader> tempList;
+
+    if(!updateAvailableSongs(tempList, playList))
+        return false;
+
+    for(SongHeader &header : tempList){
+        list.push_back(header.name);
     }
 
     return true;
@@ -236,7 +291,8 @@ bool MySql::removeSong(const SongHeader &header){
             return false;
         }
     }else if(!header.name.isEmpty() && header.size > 0){
-        QString qyer = QString("DELETE from songs where name='%0' and size=%1").arg(header.name).arg(header.size);
+        QString qyer = QString("DELETE from songs where name='%0'"
+                               " and size=%1").arg(header.name).arg(header.size);
         if(!qyery->exec(qyer)){
             sqlErrorLog(qyer);
             return false;
@@ -248,7 +304,11 @@ bool MySql::removeSong(const SongHeader &header){
 }
 
 bool MySql::addPlayList(const QString &newPlayList, const QString& desc){
-    QString qyer = QString("INSERT INTO playlists(name, description) VALUES('%0', '%1')").arg(newPlayList, desc);
+    if(newPlayList == ALL_SONGS_LIST)
+        return false;
+
+    QString qyer = QString("INSERT INTO playlists(name, description)"
+                           " VALUES('%0', '%1')").arg(newPlayList, desc);
     if(!qyery->exec(qyer)){
         sqlErrorLog(qyer);
         return false;
@@ -257,15 +317,20 @@ bool MySql::addPlayList(const QString &newPlayList, const QString& desc){
 }
 
 bool MySql::addToPlayList(const SongHeader &header, const QString &newPlaylist){
+    if(newPlaylist == ALL_SONGS_LIST)
+        return false;
+
     if(header.id > -1){
-        QString qyer = QString("INSERT INTO playlistsdata(song, playlist) VALUES(%0,'%1')").arg(header.id).arg(newPlaylist);
+        QString qyer = QString("INSERT INTO playlistsdata(song, playlist)"
+                               " VALUES(%0,'%1')").arg(header.id).arg(newPlaylist);
         if(!qyery->exec(qyer)){
             sqlErrorLog(qyer);
             return false;
         }
     }else if(!header.name.isEmpty() && header.size > 0){
         QString qyer = QString("INSERT INTO playlistsdata(song, playlist) "
-                               "VALUES((SELECT id from songs where name='%0' and size='%1'),'%2')")
+                               "VALUES((SELECT id from songs where name='%0'"
+                               " and size='%1'),'%2')")
                 .arg(header.name).arg(header.size).arg(newPlaylist);
 
         if(!qyery->exec(qyer)){
@@ -279,6 +344,9 @@ bool MySql::addToPlayList(const SongHeader &header, const QString &newPlaylist){
 }
 
 bool MySql::removeFromPlayList(const SongHeader &header, const QString &playList){
+    if(playList == ALL_SONGS_LIST)
+        return false;
+
     if(header.id > -1){
         QString qyer = QString("DELETE from playlistsdata where song=%0 and playlist='%1'").arg(header.id).arg(playList);
         if(!qyery->exec(qyer)){
@@ -287,7 +355,8 @@ bool MySql::removeFromPlayList(const SongHeader &header, const QString &playList
         }
     }else if(!header.name.isEmpty() && header.size > 0){
         QString qyer = QString("DELETE from playlistsdata where "
-                               "song=(SELECT id from songs where name='%0' and size='%1') and playlist='%2'")
+                               "song=(SELECT id from songs where name='%0' and size='%1')"
+                               " and playlist='%2'")
                 .arg(header.name).arg(header.size).arg(playList);
 
         if(!qyery->exec(qyer)){
@@ -301,6 +370,9 @@ bool MySql::removeFromPlayList(const SongHeader &header, const QString &playList
 }
 
 bool MySql::removePlayList(const QString &playList){
+    if(playList == ALL_SONGS_LIST)
+        return false;
+
     QString qyer = QString("DELETE from playlists where name='%0'").arg(playList);
     if(!qyery->exec(qyer)){
         sqlErrorLog(qyer);
@@ -309,7 +381,7 @@ bool MySql::removePlayList(const QString &playList){
     return true;
 }
 
-bool MySql::getPlayLists(QStringList &list){
+bool MySql::getPlayLists(QStringList &list)const{
     QString qyer = QString("SELECT name from playlists");
     if(!qyery->exec(qyer)){
         sqlErrorLog(qyer);
@@ -318,11 +390,34 @@ bool MySql::getPlayLists(QStringList &list){
 
     list.clear();
 
+    list.push_back(ALL_SONGS_LIST);
+
     while(qyery->next()){
         list.push_back(qyery->value(0).toString());
     }
 
     return true;
+}
+
+int MySql::getSongId(const QString &name){
+    if(name.isEmpty()){
+        return -1;
+    }
+
+    QString qyer = QString("SELECT id from songs where name='%0'").arg(name);
+    if(!qyery->exec(qyer)){
+        return -1;
+    }
+
+    if(qyery->size() != 1){
+        throw DataBaseError();
+    }
+
+    if(!qyery->next()){
+        return -1;
+    }
+
+    return qyery->value(0).toInt();;
 }
 
 void MySql::clear(){
